Refactor releases code into the action
